requirements.txt
amqp==1.4.9
anyjson==0.3.3
arrow==0.12.1
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.7.0
backports.functools-lru-cache==1.5
billiard==3.3.0.23
bleach==2.1.3
boto==2.49.0
celery==3.1.24
certifi==2018.4.16
cffi==1.11.5
characteristic==14.3.0
chardet==3.0.4
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.2.2
cssselect==1.0.3
dj-database-url==0.5.0
Django==1.11.15
django-appconf==1.0.2
django-compat==1.0.15
django-compressor==2.2
django-debug-toolbar==1.9.1
django-extensions==2.0.7
django-hijack==2.1.7
django-multiselectfield==0.1.8
django-polymorphic==1.3
django-postgrespool==0.3.1
django-redis==4.9.0
django-rest-hooks==1.5.0
django-storages==1.6.6
django-test-without-migrations==0.6
django-widget-tweaks==1.4.2
djangorestframework==3.8.2
docutils==0.14
elasticsearch==6.2.0
enum34==1.1.6
facebookads==2.11.4
factory-boy==2.11.1
fake-factory==0.5.7
Faker==0.9.0
funcsigs==1.0.2
gunicorn==19.9.0
html5lib==1.0.1
httplib2==0.11.3
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipaddress==1.0.22
itsdangerous==0.24
Jinja2==2.10
keen==0.5.1
kombu==3.0.37
lockfile==0.12.2
loxun==2.0
lxml==4.2.1
Markdown==2.6.11
MarkupSafe==0.23
measurement==1.8.0
mock==2.0.0
munch==2.3.2
MySQL-python==1.2.5
ndg-httpsclient==0.5.0
newrelic==4.2.0.100
oauth2client==4.1.2
oauthlib==2.0.7
ordereddict==1.1
pbr==4.0.2
phonenumbers==8.9.10
Pillow==5.2.0
ply==3.11
psycopg2==2.7.4
pusher==2.0.1
pyactiveresource==2.1.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pycryptodome==3.6.1
pyOpenSSL==17.5.0
python-dateutil==2.7.2
pytz==2018.4
PyYAML==3.13
queuelib==1.5.0
raven==6.9.0
rcssmin==1.0.6
redis==2.10.6
reportlab==3.5.4
requests==2.19.1
requests-mock==1.5.2
requests-oauthlib==1.0.0
rjsmin==1.0.12
rsa==3.4.2
service-identity==17.0.0
ShopifyAPI==3.1.0
simplejson==3.16.0
six==1.10.0
slimit==0.8.1
SQLAlchemy==1.2.10
sqlparse==0.2.4
stripe==1.84.2
sympy==0.7.6.1
tblib==1.3.2
text-unidecode==1.2
tld==0.7.10
tqdm==4.24.0
Twisted==18.4.0
typing==3.6.4
Unidecode==1.0.22
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
w3lib==1.19.0
webencodings==0.5.1
Werkzeug==0.14.1
whitenoise==3.3.1
WooCommerce==1.2.1
zope.interface==4.5.0
